+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- getAuthenticatedUser endpoint
+
 ## [2.173.2] - 2025-07-31
 
 ## [2.173.2] - 2025-07-28